competencia-02/requirements.txt
alembic==1.17.0
colorlog==6.10.1
dotenv==0.9.9
duckdb==1.4.1
greenlet==3.2.4
lightgbm==4.6.0
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.3.4
optuna==4.5.0
packaging==25.0
pandas==2.3.3
polars==1.34.0
polars-runtime-32==1.34.0
pyarrow==22.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
scipy==1.16.2
six==1.17.0
SQLAlchemy==2.0.44
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
